SlowQueries Timezone issue fix
timestamp is UTC but this was converted to local time in .NET (local time of agent which might be different to SQL instance - the conversion back is incorrect in this case).   The times now come back from the agent in correct UTC format and the handling of local to UTC time has been removed.