Added support for custom performance counter collection
Custom os performance counters could previously be configured by editing an XML file but this was limited to counters available in sys_dm_os_performance_counters.  It's now possible to add custom counters by creating a stored proc in the DB that the service connects to to return the required custom counters in the expected format.  e.g. Application metrics etc.