Log shipping reporting improvement
Recently created DBs were excluded until they were >1 day old.  These will now be detected as log shipping DBs with a status of N/A.  The status will display as N/A (New Database).  This makes it more obvious that we have detected it as a log shipping DB and it's excluded because it's recently created.