Bug fix for backups on availability groups
If the UTC Offset is different between servers in an AG we will get duplicate rows and it might appear that we don't have a backup.  Removed grouping by UTCOffset to fix.
Also, in this case the UTCOffset was only different due to a rare case where the dates generated between GETDATE() and GETUTCDATE() were slightly different spanning a minute interval.  Fixed by calculating the diff in seconds instead and rounding to the nearest minute.