Fix name mismatch with Databases collection
Collection reference was "Database" instead of "Databases" which is the name of the table.   Errors could be logged as "Databases" which doesn't match the collection reference. Also, stored proc is named "Database_Upd" instead of Databases_Up.
Some refactoring to fix this inconsistency and allow for generic handling of the "Databases" collection.