Fix issue with ServerExtraProperties collection for certain versions of SQL
There is a check to see if the instant_file_initialization_enabled column exists in the sys.dm_server_services DMV.  On some versions of SQL Server selecting a column that doesn't exist from a table that does exist in a branch of code that *doesn't execute* will result in an error.  e.g. 2008 R2 SP2
Dynamic SQL fixes this issue. #74