JobHistory - Exclude In progress
Exclude run_status 4 (in progress) from job history collection.  These rows can be produced by CDC.  It's incorrect to sum the durations for in progress rows and doing so can result in an overflow exception. Removing these from the collection will fix the issue and reduce the amount of data to collect/store for job history.  #147