Object Execution Stats Fix
Fix 'An item with the same key has already been added' error in the object execution chart.  This occurs when the service account doesn't have permission to collect the object name - resulting in objects with the same blank string as the name.
Replaced blank string with the object_id.
Changed sort order to further reduce the change of the error occurring.
Refactoring variable name. #158