Add testing in GitHub Action
Deploy DBA Dash to local SQL instance on the GitHub runner.
Run some tests using Pester to validate that we have data in the tables we expect to have data.
Check table counts
Check that there are no errors in the CollectionErrorLog table.
Output the content of the Logs folder and check for errors