Fix autosize issue with slow queries
Columns are no longer auto-sized on refresh, which can cause the column to grow too large if grouped by text.  Scrollbars don't appear if the first column width is larger than the width of the control. Reset to half the width of control if this situation is detected due to user resizing.