Drill down bug fix
Fix drill downs on Checks node in the tree at instance and root level.
Ensure tab load suppression does not remain enabled in the event that we can't find an instance on drill down.
Display an error message if an instance isn't found for some reason.
Move FindInstance to SQLTreeItem class and improve logic.
#472