Improve tree performance
Improve the performance of expanding an instance in the tree.
Databases have been moved to a Databases folder to avoid the need to enumerate the list of databases on expanding the instance.
AddDatabaseFolders is now done on expand.  This significantly improves performance where there is a large number of databases.