Ensure Deserialize method works when config is encrypted.
Fixes error connecting to repository database that occurs when:
You are running the GUI from the service installation folder. The config file is encrypted. You haven't modified the connections to create the ConnectionStrings.json file in the AppData folder. The error occurs even if you have password to decrypt config.