Upgrades - Get service name from GetWmiObject
It might not be possible to get the service name from the config file if the config has been encrypted.
Add a Stop-Process for DBADashService.  We should have a graceful shutdown with Stop-Service but this should ensure it's not running.  It might be running as a console app rather than installed as a service.