Avoid querying sys.dm_db_mirroring_auto_page_repair unless mirroring is used
This DMV has a bug where it can generate a dump if there are databases in a RESTORING state and the query is not run as sysadmin.  Also, the query can be blocked when databases are in a RESTORING state.  We should avoid running this DMV unless database mirroring is actually been used.
#682