Option to force data purge to run for non-partitioned tables
Data purge only runs once per day for non-partitioned tables.  With the Force parameter, the purge can run immediately regardless of the last run date.  This is useful for testing the purge process.  Also, it makes sense to run with @Force=1 if a user is manually initiating a purge from the GUI.