Ensure ConnectionID is populated for new connections
The ConnectionID is now populated from the Data Source in the connection string if it's not possible to connect to the data source at the time of adding a connection.
It's important that the ConnectionID is populated as it's used to report on offline instances and it's required for messaging to work.