Option to consider full backup with diff backup thresholds
Option to ignore the diff backup threshold if the full backup meets the criteria for the diff threshold.  e.g. Diff backups might be skipped on days when the full backup runs and you would need to extend the threshold to account for this.  If the threshold is extended it might take longer to identify an issue.  #1169