Plan collection messaging
In the running queries tab, the View Plan link provides an option to download a query plan.  If the plan is not in the repository database, this link is changed to Find Plan, and a plan collection query is provided.
Instead of providing a script, get the query plan using the plan handle directly from the monitored instance using the messaging feature (if enabled).  The Collection script can still be accessed on the SQL/Plan Handle and Query/Plan hash columns.