Query Store top queries fix
Fix error that occurs if there is a database with NUMERIC_ROUNDABORT ON.

```
Msg 1934, Level 16, State 1, Line 21
SELECT failed because the following SET options have incorrect settings: 'NUMERIC_ROUNDABORT'. Verify that SET options are correct for use with indexed views and/or indexes on computed columns and/or filtered indexes and/or query notifications and/or XML data type methods and/or spatial index operations.
```

Better handling for non-success responses.
#1350