Fix parameter was not supplied bug on Memory tab
The user is allowed to interact with the aggregate filter for the memory clerk line chart before a memory clerk is selected which causes the error.
Updated ChartView type when control is created so visibility of the control is set correctly.  Added a validation check for ShowMemoryUsageForClerk which should no longer be triggered without a selected clerk. #1378