Custom Proc Execution
Allow custom proc execution directly on monitored instances via the Messaging system.
Proc execution is limited to those explicitly specified in the config and those deployed to the monitored instance that the service account has permissions to execute.
Proc execution is also limited from the GUI to users with the Messaging + CustomTools roles, or db_owner.
Additional customization is available in the GUI.