Summary Page - Test summary splitter distance
The splitter distance is adjusted automatically to show the full test summary, but on smaller displays this doesn't leave enough room for the second grid.  This can be mitigated by using focused view.
Updated to restrict the test summary to a max % of the screen size.  It will still auto adjust to the size it needs but it won't go beyond the configured %.  This can be changed in the Options menu to the users preference.  Default 70%.