Repository DB upgrade handling improvement
If you load the GUI after upgrading to a new version of DBA Dash you might get prompt that the version of the app is NEWER than the repository database.  If an upgrade is in progress it will now say that there is an upgrade in progress and it will wait until it completes before continuing (or the user can override this).