Improve Database Finder
Database finder has been improved, showing a dialog that includes all the matched databases.  Links in this report can be used to navigate directly to the specified database/tab.
The selected item in the tree is highlighted - useful for keeping track of the current position in the tree.

Custom reports have been improved to support the new database finder report:
* A new Navigate Tree link has been added.  (Navigate to a database or instance and select the specified tab)
* A new text picker is available.