Improve deploy process
The dacpac deployment would remove the extended property before the upgrade was complete - allowing the GUI to connect before the version is upgraded.
The property is now part of the dacpac.  IsDBUpgradeInProgress is initially set to 'Y' and updated to 'N' at the end of the post deployment script.  The deploy process also sets it to 'Y' before starting the deploy.