AzureDBElasticPool expiry
As the elastic pool is no longer built based on the pools associated with instances, we need a way to retire pools that no longer exist.  Added a ValidTo column to AzureDBElasticPool that will be set when the pool is deleted.  Updated AzureDBElasticPool_Get to exclude pools deleted over a week ago.

Update AzureDBElasticPoolResourceStats_Upd.sql