Query Store context bug fix
When setting context, the object name is assigned to the object name textbox when the tree type is instance or database.  This expects to filter on the name of a stored procedure - not the database or instance name.  Bug introduced in #1598.
Added a check on the tree type.  Updated object execution tab to pass the correct tree type.
Allow Query store tab to be visible at SP level.